+++ src/main/java/kitchenpos/application/dto/response/ProductResponse.java
@@ -1,7 +1,7 @@
 package kitchenpos.application.dto.response;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 
 public class ProductResponse {
