+++ src/main/java/kitchenpos/domain/menu/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.domain.menu;
 
 import java.math.BigDecimal;
 import java.util.Objects;
