+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -8,7 +8,7 @@ import java.util.List;
 import kitchenpos.application.dto.request.ProductCommand;
 import kitchenpos.application.dto.response.ProductResponse;
 import kitchenpos.common.DataClearExtension;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.menu.Product;
 import kitchenpos.domain.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
